Initial implemention for statuses/user_timeline wrapper
This also create the app's api endpoint to access statuses/user_timeline data.
Hitting this endpoint returns a list of tweets containing the text,
user_id, the tweet id, and the created_at date. The first time we query
the API, we cache the results in redis, which lives for about 60
seconds. When the cache expires, we hit the API again. When new statues
come through, we append the previous tweets to the new ones, thus
keeping a historical record of tweets.

This requeires A LOT of refactoring. Eventually I'd like to move the
HTTP calls out into a separate library, create a wrapper for redis that
handles serialization, deserialization, etc., and move a lot of
nonessential work to a background job like writing stuff to redis.